chore: 🤖 MongoDB 미사용으로 인한 파일 제거, 패키지명과 클래스명 수정
