--- backend/joalarm/src/main/java/com/cupid/joalarm/mon/MonTest.java
@@ -1,26 +1,0 @@
-package com.cupid.joalarm.mon;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.springframework.data.mongodb.core.mapping.Document;
-
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-@Document(collection = "MonTest")
-public class MonTest {
-
-    @Id
-    private String testId;
-
-    private String title;
-
-    private String content;
-}
