--- backend/joalarm/src/main/java/com/cupid/joalarm/mon/MonTestDto.java
@@ -1,18 +1,0 @@
-package com.cupid.joalarm.mon;
-
-import lombok.*;
-
-@Getter
-@Setter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class MonTestDto {
-
-    private String testId;
-
-    private String title;
-
-    private String content;
-
-}
