--- backend/joalarm/src/main/java/com/cupid/joalarm/mon/MonTestRepository.java
@@ -1,7 +1,0 @@
-package com.cupid.joalarm.mon;
-
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-public interface MonTestRepository extends MongoRepository<MonTest, String> {
-    public MonTest findByTitle(String title);
-}
